Split handlers into handlers and utils.
