+++ tests/utils/test_process.py
@@ -1,27 +1,9 @@
 import logging
 import pytest
-from imageatm.handlers.utils import save_json, load_json, run_cmd
-from pathlib import Path
+from imageatm.utils.process import run_cmd
 
 
-TEST_TARGET_FILE = Path('./tests/data/test_samples/test_target_file.json').resolve()
-
-
-class TestUtils(object):
-    def test_save_json(self):
-        data = [
-            {'image_id': 'helmet_1.jpg', 'label': 'left'},
-            {'image_id': 'helmet_2.jpg', 'label': 'left'},
-            {'image_id': 'image_png.png', 'label': 'right'},
-        ]
-
-        target_file = TEST_TARGET_FILE
-        assert target_file.exists() is False
-        save_json(data, target_file)
-        assert target_file.exists()
-        assert load_json(target_file) == data
-        target_file.unlink()
-
+class TestProcess(object):
     def test_run_cmd_1(self, mocker):
         mp_debug = mocker.patch('logging.Logger.debug')
         mp_info = mocker.patch('logging.Logger.info')
